deriv bots python/requirements.txt
websocket-client
numpy
scikit-learn
requests
python-dotenv
pyTelegramBotAPI
matplotlib
pandas
Flask
selenium
selenium-wire
webdriver-manager
